Adding observations table (need to make relationships urgent)

+++ src/main/java/tk/soylorenzo/dao/ObservacionesDAO.java
@@ -1,9 +1,8 @@
 package tk.soylorenzo.dao;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-
 import tk.soylorenzo.models.Observaciones;
 
-public interface ObservacionesDAO extends JpaRepository<Observaciones, Integer>{
+public interface ObservacionesDAO extends JpaRepository<Observaciones, Long>{
 	
 }
