+++ src/main/java/tk/soylorenzo/dao/PersonaDAO.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import tk.soylorenzo.models.Persona;
 
-public interface PersonaDAO extends JpaRepository<Persona, Integer>{
+public interface PersonaDAO extends JpaRepository<Persona, Long>{
 	
 }
